autogen/requirements.txt
aioconsole==0.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
asyncio-atexit==1.0.1
attrs==24.2.0
autogen-agentchat==0.4.0.dev11
autogen-core==0.4.0.dev11
autogen-ext==0.4.0.dev11
certifi==2024.8.30
charset-normalizer==3.4.0
Deprecated==1.2.15
distro==1.9.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.4.0
jiter==0.8.2
jsonref==1.1.0
multidict==6.1.0
openai==1.57.2
opentelemetry-api==1.27.0
pillow==11.0.0
propcache==0.2.1
protobuf==4.25.5
pydantic==2.10.3
pydantic_core==2.27.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
tavily-python==0.5.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
